requirements.txt
flask==3.0.0
requests==2.31.0
pydub==0.25.1
audioop-lts>=0.2.1
python-dotenv==1.0.0
werkzeug==3.0.0
psycopg2-binary>=2.9.9
gunicorn==21.2.0
google-generativeai>=0.8.0
flask-cors==4.0.0
pyjwt==2.8.0
bcrypt==4.1.2
razorpay==1.4.1